separate SpringBoneLogic.cs
