+++ Assets/UniVRM-1.0/Components/SpringBone/SpringBoneLogic.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: de6086bc4a9b4434b8cabc6e9e81eb4d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
